requirements.txt
apache-airflow==3.1.6
apache-airflow-providers-amazon==8.1.0
apache-airflow-providers-http==4.11.0
requests==2.32.3
pandas==2.2.0
boto3==1.34.0
apache-airflow-providers-snowflake>=6.7.0
snowflake-connector-python>=2.8.0